change all cpp files to c files
