+++ CProgremDesign/chapter2/2-10.c
