+++ CProgremDesign/chapter2/2-1A.c
