+++ CProgremDesign/chapter2/2-1B.c
