+++ CProgremDesign/chapter2/2-5.c
