+++ CProgremDesign/chapter2/2-6.c
