+++ CProgremDesign/chapter2/2-7.c
