+++ CProgremDesign/chapter2/2-8.c
